Updated to follow the new bundle syntax and remove the Bundle namespace
This commit change the namespace from Bundle\FOS\UserBundle to FOS\UserBundle
as the Bundle part is now useless (and would mean naming the bundle
BundleFOSUserBundle).
It will require you to change your submodules and to register the FOS
namespace in the autoloader.

It also fixes the FOSUserBundle class to use the new BundleInterface.