Added a mailer able to use twig blocks for the each part of the message
It expects the template to define 3 blocks: subject, body_text and
body_html. The existing templates have been refactored to define these
blocks, but maintaining BC with the other mailer implementation and don't
provide the html body.
The new mailer is not used by default to maintain BC as it requires using
Twig for the templates. You can switch to it with this config:

fos_user:
    service:
        mailer: fos_user.mailer.twig_swift