Removed the option to switch to a different templating engine
This option was introduced in the early days of the Symfony development
when Twig was not yet the recommended engine, to let users remove the
dependency on Twig at the expense of rewriting all templates for their
engine.
Making the code ugly and inconsistent with the Symfony ecosystem to
implement a feature which is barely usable is not worth it.