use relative refs
Often, people report issues for FOSUserBundle `1.3.x` because they
followed the documentation for `2.x`. Currently, the readme links to
the master version of the docs regardless of the branch you are
browsing. Using relative links so that users get the right version of
the docs should mitigate/reduce those issues.