Remove more fields from the User class
These fields are used by advanced features of the security system, which
are not ready out of the box in the bundle anyway. The provided fields
was the cause of complaints from 2 sides:
- people complaining about useless fields in the DB when they don't use
  the feature
- people complaining about the implementation not fitting their business
  needs.
The bundle does not provide the storage for these properties anymore.
Projects needing to use these features should add the fields and the
implementation they need in their child User class instead.