Remove the group feature
This feature has been deprecated in 2.2 because there are much better
alternatives than using groups to assign roles to users through the
group (which was not even working fine with the default serialization of
users). The role hierarchy of Symfony solves basic use cases (a
non-dynamic hierarchy of roles) in a simpler way. And for more complex
rules, the official solution is to use custom voters instead of voting
on roles and building a complex way to assign roles.