Fix a deprecation about groups being triggered when loading ORM metadata
When using the `getAllMetadata` API of the Doctrine ORM, it will load
the abstract Group class due to being a MappedSuperClass in the mapping.
This was triggering the deprecation warning about using groups, even
though they were not actually used.
Given that this class is abstract, the only way to actually use it
involves extending it, which gets a deprecation warning triggered by the
DebugClassLoader.