Fix deprecation warning for UserInterface on Symfony 4
The deprecated annotation removed in 2.2.3 had a purpose: deprecated
interfaces are allowed to extend deprecated interfaces of another
package without triggering a warning in DebugClassLoader.
However, applying it on the public UserInterface was the mistake.
Instead of having an internal interface that is extended by the
conditional declaration of the public interface, the implementation is
now reversed. An internal CompatUserInterface is now declared
conditionally to handle the BC layer, and the public UserInterface
extends it.
As our UserInterface and the CompatUserInterface are in the same
package, UserInterface is allowed to extend CompatUserInterface without
warning even when it is marked as deprecated, making both
DebugClassLoader and downstream static analyzers happy.